refactor: pin 리팩토링
1. Pin -> NomalPin으로 변경
2. Pin 인터페이스 생성
3. 패키지 변경
