+++ src/main/java/bowling/domain/pin/NormalPin.java
@@ -1,27 +1,28 @@
-package bowling.domain;
+package bowling.domain.pin;
 
 import java.util.Objects;
 
-public class Pin {
+public class NormalPin implements Pin {
     public static final int MAX_PIN_COUNT = 10;
     public static final int MIN_PIN_COUNT = 0;
 
     private int pin;
 
-    private Pin(int countOfDownPin) {
+    private NormalPin(int countOfDownPin) {
         this.pin = countOfDownPin;
     }
 
-    public static Pin from(int countOfDownPin) {
+    public static NormalPin from(int countOfDownPin) {
         if (countOfDownPin > MAX_PIN_COUNT) {
             throw new IllegalArgumentException("핀은 최대 10개까지 쓰러뜨릴 수 있습니다.");
         }
         if (countOfDownPin < MIN_PIN_COUNT) {
             throw new IllegalArgumentException("핀은 최소 0개까지 쓰러뜨릴 수 있습니다.");
         }
-        return new Pin(countOfDownPin);
+        return new NormalPin(countOfDownPin);
     }
 
+    @Override
     public int getPin() {
         return pin;
     }
@@ -30,8 +31,8 @@ public class Pin {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        Pin pin1 = (Pin) o;
-        return pin == pin1.pin;
+        NormalPin normalPin1 = (NormalPin) o;
+        return pin == normalPin1.pin;
     }
 
     @Override
