+++ src/main/java/bowling/domain/pin/Pin.java
@@ -1,0 +1,6 @@
+package bowling.domain.pin;
+
+public interface Pin {
+
+    int getPin();
+}
